Initial Commit - Empty Hello World
This repository is for the AWS RoboMaker 'Hello World' Twitch episode
where we demonstrate how you can code a simple robot, and build,
simulate and test it using AWS RoboMaker.

It is based off the aws-robomaker-sample-application-hello-world
repository with most of the core robot code removed and only the
boilerplate code remaining. We will be adding the code to make the robot
work during the Twitch stream.